techpack: audio: Properly allocate adm_session
Compiling the kernel with LTO_CLANG exposes an issue with the high
quality, programmatically sound DSP additions masterfully crafted
in Shenzhen. OnePlus has so heavily optimized their code, that LTO
can't possibly push it's performance any further. As a result,
their code is optimized out of existence and attempting to utilize
the device's microphone results in an immediate kernel panic
with following call trace:

[   51.897451] Call trace:
[   51.897457]  msm_pcm_routing_reg_phy_stream+0x4b0/0xabc
[   51.897463]  msm_pcm_prepare.106968+0x6f8/0x7dc
[   51.897470]  soc_pcm_prepare+0xa4/0x3f8
[   51.897476]  dpcm_fe_dai_prepare+0x1cc/0x388
[   51.897482]  snd_pcm_common_ioctl+0x1d68/0x4c84
[   51.897487]  snd_pcm_ioctl_compat+0x320/0x268c
[   51.897494]  compat_SyS_ioctl+0x150/0x260
[   51.897501]  el0_svc_naked+0x34/0x38
[   51.897508] Code: b94073e9 8b29c908 2a0003e4 aa1503fa (b905d115)

In all seriousness, improper adm_session allocation is resulting in
it's inaccessibility to msm_pcm_routing_reg_phy_stream. Remedy this
this by utilizing kzalloc so that functions calling adm_session have
an addressable location in RAM to find it.

Thanks to @YaroST12 for volunteering his time, and superior PC to
crank out about a dozen LTO builds over the course of a couple hours
to help me debug this.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: Yaroslav Furman <yaro330@gmail.com>