Partially revert "drm: msm: Add separate sysfs node for FOD dimming alpha"
When the aforementioned commit was introduced, ROMs that did not use it had
slower FP as usual, sometimes as slow as 2-3x. However, this commit is needed
to ensure compatibility of FOD support on maximum ROMs. Partially revert it and
add back the condition for 'op_dimlayer_bl_enable' to ensure that normal speeds
are retained.

This will likely break DC dimming on custom ROMs but that's a sacrifice
I'm willing to make as having working FOD is a higher priority until all ROMs
unify their implementations.

Signed-off-by: idkwhoiam322 <idkwhoiam322@raphielgang.org>